Save Last Selection properly remember last Section and Link selected

+++ src/menu.cpp
@@ -45,11 +45,9 @@ Menu::Menu(GMenu2X *gmenu2x) {
 	if ((dirp = opendir("sections")) == NULL) return;
 
 	while ((dptr = readdir(dirp))) {
-		if (dptr->d_name[0]=='.') continue;
-		filepath = (string)"sections/"+dptr->d_name;
-		int statRet = stat(filepath.c_str(), &st);
-		if (!S_ISDIR(st.st_mode)) continue;
-		if (statRet != -1) {
+		if (dptr->d_name[0] == '.') continue;
+		string filepath = (string)"sections/" + dptr->d_name;
+		if (dir_exists(filepath)) {
 			sections.push_back((string)dptr->d_name);
 			linklist ll;
 			links.push_back(ll);
@@ -75,14 +73,14 @@ uint Menu::firstDispRow() {
 
 void Menu::loadIcons() {
 	//reload section icons
-	for (uint i=0; i<sections.size(); i++) {
-		string sectionIcon = "sections/"+sections[i]+".png";
+	for (uint i = 0; i < sections.size(); i++) {
+		string sectionIcon = "sections/" + sections[i] + ".png";
 		if (!gmenu2x->sc.getSkinFilePath(sectionIcon).empty())
-			gmenu2x->sc.add("skin:"+sectionIcon);
+			gmenu2x->sc.add("skin:" + sectionIcon);
 
 		//check link's icons
 		string linkIcon;
-		for (uint x=0; x<sectionLinks(i)->size(); x++) {
+		for (uint x = 0; x < sectionLinks(i)->size(); x++) {
 			linkIcon = sectionLinks(i)->at(x)->getIcon();
 			sectionLinks(i)->at(x)->updateSurfaces();
 			LinkApp *linkapp = dynamic_cast<LinkApp*>(sectionLinks(i)->at(x));
@@ -92,7 +90,7 @@ void Menu::loadIcons() {
 				linkapp->searchManual();
 			}
 
-			if (linkIcon.substr(0,5)=="skin:") {
+			if (linkIcon.substr(0,5) == "skin:") {
 				linkIcon = gmenu2x->sc.getSkinFilePath(linkIcon.substr(5,linkIcon.length()));
 				if (linkapp != NULL && !fileExists(linkIcon))
 					linkapp->searchIcon();
@@ -110,17 +108,19 @@ void Menu::loadIcons() {
    SECTION MANAGEMENT
   ====================================*/
 void Menu::freeLinks() {
-	for (vector<linklist>::iterator section = links.begin(); section<links.end(); section++)
-		for (linklist::iterator link = section->begin(); link<section->end(); link++)
+	for (vector<linklist>::iterator section = links.begin(); section < links.end(); section++)
+		for (linklist::iterator link = section->begin(); link < section->end(); link++)
 			delete *link;
 }
 
 linklist *Menu::sectionLinks(int i) {
-	if (i<0 || i>(int)links.size())
+	if (i < 0 || i > (int)links.size()) {
 		i = selSectionIndex();
+	}
 
-	if (i<0 || i>(int)links.size())
+	if (i < 0 || i > (int)links.size()) {
 		return NULL;
+	}
 
 	return &links[i];
 }
@@ -171,8 +171,11 @@ void Menu::setSectionIndex(int i) {
 }
 
 string Menu::sectionPath(int section) {
-	if (section<0 || section>(int)sections.size()) section = iSection;
-	return "sections/"+sections[section]+"/";
+	if (section < 0 || section > (int)sections.size()) {
+		section = iSection;
+	}
+
+	return "sections/" + sections[section] + "/";
 }
 
 /*====================================
@@ -187,7 +190,7 @@ bool Menu::addActionLink(uint section, const string &title, LinkRunAction action
 	linkact->setSize(gmenu2x->linksRect.w, gmenu2x->skinConfInt["linkItemHeight"]);
 	linkact->setTitle(title);
 	linkact->setDescription(description);
-	if (gmenu2x->sc.exists(icon) || (icon.substr(0,5)=="skin:" && !gmenu2x->sc.getSkinFilePath(icon.substr(5,icon.length())).empty()) || fileExists(icon))
+	if (gmenu2x->sc.exists(icon) || (icon.substr(0,5) == "skin:" && !gmenu2x->sc.getSkinFilePath(icon.substr(5, icon.length())).empty()) || fileExists(icon))
 	linkact->setIcon(icon);
 
 	sectionLinks(section)->push_back(linkact);
@@ -197,7 +200,7 @@ bool Menu::addActionLink(uint section, const string &title, LinkRunAction action
 bool Menu::addLink(string path, string file, string section) {
 	if (section=="")
 		section = selSection();
-	else if (find(sections.begin(),sections.end(),section)==sections.end()) {
+	else if (find(sections.begin(), sections.end(), section) == sections.end()) {
 		//section directory doesn't exists
 		if (!addSection(section))
 			return false;
@@ -209,21 +212,21 @@ bool Menu::addLink(string path, string file, string section) {
 	//strip the extension from the filename
 	string title = file;
 	string::size_type pos = title.rfind(".");
-	if (pos!=string::npos && pos>0) {
+	if (pos != string::npos && pos > 0) {
 		string ext = title.substr(pos, title.length());
 		transform(ext.begin(), ext.end(), ext.begin(), (int(*)(int)) tolower);
 		if (ext == ".gpu" || ext == ".gpe") wrapper = false;
 		title = title.substr(0, pos);
 	}
 
-	string linkpath = "sections/"+section+"/"+title;
+	string linkpath = "sections/" + section + "/" + title;
 	int x=2;
 	while (fileExists(linkpath)) {
 		stringstream ss;
 		linkpath = "";
 		ss << x;
 		ss >> linkpath;
-		linkpath = "sections/"+section+"/"+title+linkpath;
+		linkpath = "sections/" + section + "/" + title + linkpath;
 		x++;
 	}
 
@@ -302,8 +305,8 @@ bool Menu::addLink(string path, string file, string section) {
 }
 
 bool Menu::addSection(const string &sectionName) {
-	string sectiondir = "sections/"+sectionName;
-	if (mkdir(sectiondir.c_str(),0777)==0) {
+	string sectiondir = "sections/" + sectionName;
+	if (mkdir(sectiondir.c_str(), 0777) == 0) {
 		sections.push_back(sectionName);
 		linklist ll;
 		links.push_back(ll);
@@ -317,8 +320,7 @@ void Menu::deleteSelectedLink() {
 
 	INFO("Deleting link '%s'", selLink()->getTitle().c_str());
 
-	if (selLinkApp()!=NULL)
-		unlink(selLinkApp()->getFile().c_str());
+	if (selLinkApp() != NULL) unlink(selLinkApp()->getFile().c_str());
 	sectionLinks()->erase( sectionLinks()->begin() + selLinkIndex() );
 	setLinkIndex(selLinkIndex());
 
@@ -338,19 +340,19 @@ void Menu::deleteSelectedLink() {
 void Menu::deleteSelectedSection() {
 	INFO("Deleting section '%s'", selSection().c_str());
 
-	gmenu2x->sc.del("sections/"+selSection()+".png");
-	links.erase( links.begin()+selSectionIndex() );
-	sections.erase( sections.begin()+selSectionIndex() );
+	gmenu2x->sc.del("sections/" + selSection() + ".png");
+	links.erase( links.begin() + selSectionIndex() );
+	sections.erase( sections.begin() + selSectionIndex() );
 	setSectionIndex(0); //reload sections
 }
 
 bool Menu::linkChangeSection(uint linkIndex, uint oldSectionIndex, uint newSectionIndex) {
-	if (oldSectionIndex<sections.size() && newSectionIndex<sections.size() && linkIndex<sectionLinks(oldSectionIndex)->size()) {
+	if (oldSectionIndex < sections.size() && newSectionIndex < sections.size() && linkIndex < sectionLinks(oldSectionIndex)->size()) {
 		sectionLinks(newSectionIndex)->push_back( sectionLinks(oldSectionIndex)->at(linkIndex) );
-		sectionLinks(oldSectionIndex)->erase( sectionLinks(oldSectionIndex)->begin()+linkIndex );
+		sectionLinks(oldSectionIndex)->erase( sectionLinks(oldSectionIndex)->begin() + linkIndex );
 		//Select the same link in the new position
 		setSectionIndex(newSectionIndex);
-		setLinkIndex(sectionLinks(newSectionIndex)->size()-1);
+		setLinkIndex(sectionLinks(newSectionIndex)->size() - 1);
 		return true;
 	}
 	return false;
@@ -368,7 +370,7 @@ void Menu::pageUp() {
 void Menu::pageDown() {
 	// PAGEDOWN with right
 	if (iLink + gmenu2x->linkRows > sectionLinks()->size()) {
-		setLinkIndex(sectionLinks()->size()-1);
+		setLinkIndex(sectionLinks()->size() - 1);
 	} else {
 		setLinkIndex(iLink + gmenu2x->linkRows - 1);
 	}
@@ -376,24 +378,24 @@ void Menu::pageDown() {
 
 
 void Menu::linkLeft() {
-	if (iLink%gmenu2x->linkColumns == 0)
-		setLinkIndex( sectionLinks()->size()>iLink+gmenu2x->linkColumns-1 ? iLink+gmenu2x->linkColumns-1 : sectionLinks()->size()-1 );
+	if (iLink % gmenu2x->linkColumns == 0)
+		setLinkIndex(sectionLinks()->size() > iLink + gmenu2x->linkColumns - 1 ? iLink + gmenu2x->linkColumns - 1 : sectionLinks()->size() - 1 );
 	else
-		setLinkIndex(iLink-1);
+		setLinkIndex(iLink - 1);
 }
 
 void Menu::linkRight() {
-	if (iLink%gmenu2x->linkColumns == (gmenu2x->linkColumns-1) || iLink == (int)sectionLinks()->size()-1)
-		setLinkIndex(iLink-iLink%gmenu2x->linkColumns);
+	if (iLink % gmenu2x->linkColumns == (gmenu2x->linkColumns - 1) || iLink == (int)sectionLinks()->size() - 1)
+		setLinkIndex(iLink - iLink % gmenu2x->linkColumns);
 	else
-		setLinkIndex(iLink+1);
+		setLinkIndex(iLink + 1);
 }
 
 void Menu::linkUp() {
-	int l = iLink-gmenu2x->linkColumns;
-	if (l<0) {
-		uint rows = (uint)ceil(sectionLinks()->size()/(double)gmenu2x->linkColumns);
-		l = (rows*gmenu2x->linkColumns)+l;
+	int l = iLink - gmenu2x->linkColumns;
+	if (l < 0) {
+		uint rows = (uint)ceil(sectionLinks()->size() / (double)gmenu2x->linkColumns);
+		l = (rows * gmenu2x->linkColumns) + l;
 		if (l >= (int)sectionLinks()->size())
 			l -= gmenu2x->linkColumns;
 	}
@@ -418,7 +420,10 @@ int Menu::selLinkIndex() {
 }
 
 Link *Menu::selLink() {
-	if (sectionLinks()->size()==0) return NULL;
+	if (sectionLinks()->size() == 0) {
+		return NULL;
+	}
+
 	return sectionLinks()->at(iLink);
 }
 
@@ -427,15 +432,15 @@ LinkApp *Menu::selLinkApp() {
 }
 
 void Menu::setLinkIndex(int i) {
-	if (i<0)
-		i=sectionLinks()->size()-1;
-	else if (i>=(int)sectionLinks()->size())
-		i=0;
+	if (i < 0)
+		i = sectionLinks()->size() - 1;
+	else if (i >= (int)sectionLinks()->size())
+		i = 0;
 
-	if (i>=(int)(iFirstDispRow*gmenu2x->linkColumns+gmenu2x->linkColumns*gmenu2x->linkRows))
-		iFirstDispRow = i/gmenu2x->linkColumns-gmenu2x->linkRows+1;
-	else if (i<(int)(iFirstDispRow*gmenu2x->linkColumns))
-		iFirstDispRow = i/gmenu2x->linkColumns;
+	if (i >= (int)(iFirstDispRow * gmenu2x->linkCols + gmenu2x->linkCols * gmenu2x->linkRows))
+		iFirstDispRow = i / gmenu2x->linkCols - gmenu2x->linkRows + 1;
+	else if (i < (int)(iFirstDispRow * gmenu2x->linkCols))
+		iFirstDispRow = i / gmenu2x->linkCols;
 
 	iLink = i;
 }
@@ -451,15 +456,15 @@ void Menu::readLinks() {
 	struct dirent *dptr;
 	string filepath;
 
-	for (uint i=0; i<links.size(); i++) {
+	for (uint i = 0; i < links.size(); i++) {
 		links[i].clear();
 		linkfiles.clear();
 
 		if ((dirp = opendir(sectionPath(i).c_str())) == NULL) continue;
 
 		while ((dptr = readdir(dirp))) {
-			if (dptr->d_name[0]=='.') continue;
-			filepath = sectionPath(i)+dptr->d_name;
+			if (dptr->d_name[0] == '.') continue;
+			filepath = sectionPath(i) + dptr->d_name;
 			int statRet = stat(filepath.c_str(), &st);
 			if (S_ISDIR(st.st_mode)) continue;
 			if (statRet != -1) {
@@ -468,7 +473,7 @@ void Menu::readLinks() {
 		}
 
 		sort(linkfiles.begin(), linkfiles.end(),case_less());
-		for (uint x=0; x<linkfiles.size(); x++) {
+		for (uint x = 0; x < linkfiles.size(); x++) {
 			LinkApp *link = new LinkApp(gmenu2x, gmenu2x->input, linkfiles[x].c_str());
 			link->setSize(gmenu2x->linksRect.w, gmenu2x->skinConfInt["linkItemHeight"]);
 			if (link->targetExists())
